cv-build/requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
typst==0.13.2
watchdog==6.0.0
